apps/devdesign/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic-settings>=2.0.0
email-validator>=2.0
